Smart Shapes part 3

+++ src/importexport/finale/internal/finaletypesconv.h
@@ -67,7 +67,7 @@ public:
     static engraving::SlurStyleType slurStyleTypeFromShapeType(musx::dom::others::SmartShape::ShapeType shapeType);
     static engraving::DirectionV directionVFromShapeType(musx::dom::others::SmartShape::ShapeType shapeType);
     static engraving::LineType lineTypeFromShapeType(musx::dom::others::SmartShape::ShapeType shapeType);
-    static std::pair<double, double> hookHeightsFromShapeType(musx::dom::others::SmartShape::ShapeType shapeType);
+    static std::pair<int, int> hookHeightsFromShapeType(musx::dom::others::SmartShape::ShapeType shapeType);
     // unit conversion
     static double doubleFromEvpu(double evpu);
     static engraving::PointF evpuToPointF(double xEvpu, double yEvpu);
